NormalEquations has been renamed to ImagingNormalEquations. Methods unrelated to the approximated case (i.e. those implemented in GenericNormalEquations) will be removed soon.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@869 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Synthesis/synthesis/trunk/parallel/SimParallel.cc
@@ -199,7 +199,7 @@ namespace conrad
 			    }
 
 			    const vector<string> sources=parset.getStringVector("sources.names");
-			    for (int i=0; i<sources.size(); ++i)
+			    for (size_t i=0; i<sources.size(); ++i)
 			    {
 				    {
 					    ostringstream oos;
